NDS: wrap libnds exception handler instead of using a copy
